Use ZF's existing modules in Twig's file loader
This reads configured modules from the FrontController and uses these to populate the file loader.
The default module becomes the `__main__` namespace, but the other modules become literal namespaces in Twig.

This translates a bit more natural to Zend Framework's `partial` system, where module can be passed as a parameter.

```
$this->partial('partials/my_file.phtml', 'admin')
```

becomes

```
{% include('@admin/partials/my_file.twig') %}
```